@@ -10,7 +10,7 @@ description: A flutter boilerplate project created using MobX and Provider.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
 
 dependencies:
   flutter:
@@ -22,24 +22,21 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.2
 
   # The following adds the shared pref as a dependency in your application
-  shared_preferences: ^0.5.10
+  shared_preferences: ^2.0.5
 
   # A composable, Future-based library for making HTTP requests.
   http: ^0.12.2
 
-  # Flogs, a powerful logging framework
-  f_logs: ^1.2.2
-
   # MobX is a library for reactively managing the state of your applications. Use the power of
   # observables, actions, and reactions to supercharge your Dart and Flutter apps.
-  mobx: ^1.2.1+2
+  mobx: ^1.2.1+4
   flutter_mobx: ^1.1.0+2
 
   # An helper to easily exposes a value using InheritedWidget without having to write one.
-  provider: ^4.3.2+1
+  provider: ^5.0.0
 
   # String validation and sanitization for Dart. Dart 2-compatible version of validator
   validators: ^2.0.1
@@ -52,11 +49,18 @@ dependencies:
   flushbar: ^1.10.4
 
   # Material Dialog
-  material_dialog: ^0.0.9
+#  material_dialog: ^0.0.9
 
   # This package provides internationalization and localization facilities, including message
   # translation, plurals and genders, date/number formatting and parsing, and bidirectional text.
-  intl: ^0.16.1
+  intl: ^0.17.0
+
+  # An helper to easily exposes a value using InheritedWidget without having to write one.
+  sembast: ^2.4.4+1
+  path_provider: ^2.0.1
+
+  # XXTEA is a fast and secure encryption algorithm. This is a XXTEA library for Dart.
+  xxtea: ^2.0.3
 
   # Dependency injection
   inject:
@@ -74,7 +78,7 @@ dev_dependencies:
   # flutter packages get
   # flutter packages pub run flutter_launcher_icons:main
   flutter_launcher_icons: ^0.7.5
-  mobx_codegen: ^1.1.0+1
+  mobx_codegen: ^1.1.2
   json_serializable: ^3.4.1
   build_runner: ^1.10.2
   source_gen: 0.9.6
