@@ -7,7 +7,7 @@ class ErrorStore = _ErrorStore with _$ErrorStore;
 abstract class _ErrorStore with Store {
 
   // disposers
-  List<ReactionDisposer> _disposers;
+  late List<ReactionDisposer> _disposers;
 
   // constructor:---------------------------------------------------------------
   _ErrorStore() {
@@ -20,7 +20,6 @@ abstract class _ErrorStore with Store {
   @observable
   String errorMessage = '';
 
-
   // actions:-------------------------------------------------------------------
   @action
   void setErrorMessage(String message) {
@@ -29,7 +28,7 @@ abstract class _ErrorStore with Store {
 
   @action
   void reset(String value) {
-    print('calling reset');
+    print('calling reset to remove message');
     errorMessage = '';
   }
 
