@@ -8,7 +8,6 @@ import 'package:boilerplate/widgets/progress_indicator_widget.dart';
 import 'package:flushbar/flushbar_helper.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:material_dialog/material_dialog.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -19,9 +18,9 @@ class HomeScreen extends StatefulWidget {
 
 class _HomeScreenState extends State<HomeScreen> {
   //stores:---------------------------------------------------------------------
-  PostStore _postStore;
-  ThemeStore _themeStore;
-  LanguageStore _languageStore;
+  late PostStore _postStore;
+  late ThemeStore _themeStore;
+  late LanguageStore _languageStore;
 
   @override
   void initState() {
@@ -52,9 +51,9 @@ class _HomeScreenState extends State<HomeScreen> {
   }
 
   // app bar methods:-----------------------------------------------------------
-  Widget _buildAppBar() {
+  PreferredSizeWidget _buildAppBar() {
     return AppBar(
-      title: Text(AppLocalizations.of(context).translate('home_tv_posts')),
+      title: Text(AppLocalizations.of(context)!.translate('home_tv_posts')),
       actions: _buildActions(context),
     );
   }
@@ -99,7 +98,7 @@ class _HomeScreenState extends State<HomeScreen> {
   Widget _buildLanguageButton() {
     return IconButton(
       onPressed: () {
-        _buildLanguageDialog();
+        // _buildLanguageDialog();
       },
       icon: Icon(
         Icons.language,
@@ -140,7 +139,7 @@ class _HomeScreenState extends State<HomeScreen> {
           )
         : Center(
             child: Text(
-              AppLocalizations.of(context).translate('home_tv_no_post_found'),
+              AppLocalizations.of(context)!.translate('home_tv_no_post_found'),
             ),
           );
   }
@@ -183,7 +182,7 @@ class _HomeScreenState extends State<HomeScreen> {
       if (message != null && message.isNotEmpty) {
         FlushbarHelper.createError(
           message: message,
-          title: AppLocalizations.of(context).translate('home_tv_error'),
+          title: AppLocalizations.of(context)!.translate('home_tv_error'),
           duration: Duration(seconds: 3),
         )..show(context);
       }
@@ -192,58 +191,58 @@ class _HomeScreenState extends State<HomeScreen> {
     return SizedBox.shrink();
   }
 
-  _buildLanguageDialog() {
-    _showDialog<String>(
-      context: context,
-      child: MaterialDialog(
-        borderRadius: 5.0,
-        enableFullWidth: true,
-        title: Text(
-          AppLocalizations.of(context).translate('home_tv_choose_language'),
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: 16.0,
-          ),
-        ),
-        headerColor: Theme.of(context).primaryColor,
-        backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-        closeButtonColor: Colors.white,
-        enableCloseButton: true,
-        enableBackButton: false,
-        onCloseButtonClicked: () {
-          Navigator.of(context).pop();
-        },
-        children: _languageStore.supportedLanguages
-            .map(
-              (object) => ListTile(
-                dense: true,
-                contentPadding: EdgeInsets.all(0.0),
-                title: Text(
-                  object.language,
-                  style: TextStyle(
-                    color: _languageStore.locale == object.locale
-                        ? Theme.of(context).primaryColor
-                        : _themeStore.darkMode ? Colors.white : Colors.black,
-                  ),
-                ),
-                onTap: () {
-                  Navigator.of(context).pop();
-                  // change user language based on selected locale
-                  _languageStore.changeLanguage(object.locale);
-                },
-              ),
-            )
-            .toList(),
-      ),
-    );
-  }
+  // _buildLanguageDialog() {
+  //   _showDialog<String>(
+  //     context: context,
+  //     child: MaterialDialog(
+  //       borderRadius: 5.0,
+  //       enableFullWidth: true,
+  //       title: Text(
+  //         AppLocalizations.of(context)!.translate('home_tv_choose_language'),
+  //         style: TextStyle(
+  //           color: Colors.white,
+  //           fontSize: 16.0,
+  //         ),
+  //       ),
+  //       headerColor: Theme.of(context).primaryColor,
+  //       backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+  //       closeButtonColor: Colors.white,
+  //       enableCloseButton: true,
+  //       enableBackButton: false,
+  //       onCloseButtonClicked: () {
+  //         Navigator.of(context).pop();
+  //       },
+  //       children: _languageStore.supportedLanguages
+  //           .map(
+  //             (object) => ListTile(
+  //               dense: true,
+  //               contentPadding: EdgeInsets.all(0.0),
+  //               title: Text(
+  //                 object.language,
+  //                 style: TextStyle(
+  //                   color: _languageStore.locale == object.locale
+  //                       ? Theme.of(context).primaryColor
+  //                       : _themeStore.darkMode ? Colors.white : Colors.black,
+  //                 ),
+  //               ),
+  //               onTap: () {
+  //                 Navigator.of(context).pop();
+  //                 // change user language based on selected locale
+  //                 _languageStore.changeLanguage(object.locale);
+  //               },
+  //             ),
+  //           )
+  //           .toList(),
+  //     ),
+  //   );
+  // }
 
-  _showDialog<T>({BuildContext context, Widget child}) {
-    showDialog<T>(
-      context: context,
-      builder: (BuildContext context) => child,
-    ).then<void>((T value) {
-      // The value passed to Navigator.pop() or null.
-    });
-  }
+  // _showDialog<T>({required BuildContext context, required Widget child}) {
+  //   showDialog<T>(
+  //     context: context,
+  //     builder: (BuildContext context) => child,
+  //   ).then<void>((T value) {
+  //     // The value passed to Navigator.pop() or null.
+  //   });
+  // }
 }
