@@ -1,6 +1,5 @@
 import 'dart:convert';
 
-import 'package:meta/meta.dart';
 import 'package:sembast/sembast.dart';
 import 'package:xxtea/xxtea.dart';
 
@@ -11,7 +10,7 @@ class _XXTeaEncoder extends Converter<Map<String, dynamic>, String> {
 
   @override
   String convert(Map<String, dynamic> input) =>
-      xxtea.encryptToString(json.encode(input), key);
+      xxtea.encryptToString(json.encode(input), key)!;
 }
 
 class _XXTeaDecoder extends Converter<String, Map<String, dynamic>> {
@@ -21,7 +20,7 @@ class _XXTeaDecoder extends Converter<String, Map<String, dynamic>> {
 
   @override
   Map<String, dynamic> convert(String input) {
-    var result = json.decode(xxtea.decryptToString(input, key));
+    var result = json.decode(xxtea.decryptToString(input, key)!);
     if (result is Map) {
       return result.cast<String, dynamic>();
     }
@@ -32,8 +31,8 @@ class _XXTeaDecoder extends Converter<String, Map<String, dynamic>> {
 /// Simple encryption codec using xxtea
 /// It requires a password to encrypt/decrypt the data
 class _XXTeaCodec extends Codec<Map<String, dynamic>, String> {
-  _XXTeaEncoder _encoder;
-  _XXTeaDecoder _decoder;
+  late _XXTeaEncoder _encoder;
+  late _XXTeaDecoder _decoder;
 
   /// A non null [password] to use for the encryption/decryption
   _XXTeaCodec(String password) {
@@ -60,5 +59,5 @@ class _XXTeaCodec extends Codec<Map<String, dynamic>, String> {
 ///
 /// // ...your database is ready to use as encrypted
 /// ```
-SembastCodec getXXTeaCodec({@required String password}) =>
+SembastCodec getXXTeaCodec({required String password}) =>
     SembastCodec(signature: 'xxtea', codec: _XXTeaCodec(password));
