@@ -7,7 +7,7 @@ import 'package:flutter/services.dart';
 class AppLocalizations {
   // localization variables
   final Locale locale;
-  Map<String, String> localizedStrings;
+  late Map<String, String> localizedStrings;
 
   // Static member to have a simple access to the delegate from the MaterialApp
   static const LocalizationsDelegate<AppLocalizations> delegate =
@@ -18,7 +18,7 @@ class AppLocalizations {
 
   // Helper method to keep the code in the widgets concise
   // Localizations are accessed using an InheritedWidget "of" syntax
-  static AppLocalizations of(BuildContext context) {
+  static AppLocalizations? of(BuildContext context) {
     return Localizations.of<AppLocalizations>(context, AppLocalizations);
   }
 
@@ -40,7 +40,7 @@ class AppLocalizations {
 
   // This method will be called from every widget which needs a localized text
   String translate(String key) {
-    return localizedStrings[key];
+    return localizedStrings[key]!;
   }
 }
 
