@@ -6,34 +6,34 @@ class TextFieldWidget extends StatelessWidget {
   final String errorText;
   final bool isObscure;
   final bool isIcon;
-  final TextInputType inputType;
+  final TextInputType? inputType;
   final TextEditingController textController;
   final EdgeInsets padding;
   final Color hintColor;
   final Color iconColor;
-  final FocusNode focusNode;
-  final ValueChanged onFieldSubmitted;
+  final FocusNode? focusNode;
+  final ValueChanged? onFieldSubmitted;
   final ValueChanged onChanged;
   final bool autoFocus;
-  final TextInputAction inputAction;
+  final TextInputAction? inputAction;
 
   const TextFieldWidget({
-    Key key,
-    this.icon,
-    this.hint,
-    this.errorText,
-    this.isObscure = false,
+    Key? key,
+    required this.icon,
+    required this.hint,
+    required this.errorText,
+    required this.textController,
+    required this.onChanged,
+    this.inputAction,
+    this.onFieldSubmitted,
     this.inputType,
-    this.textController,
+    this.focusNode,
+    this.isObscure = false,
     this.isIcon = true,
     this.padding = const EdgeInsets.all(0),
     this.hintColor = Colors.grey,
     this.iconColor = Colors.grey,
-    this.focusNode,
-    this.onFieldSubmitted,
-    this.onChanged,
     this.autoFocus = false,
-    this.inputAction,
   }) : super(key: key);
 
   @override
@@ -54,7 +54,7 @@ class TextFieldWidget extends StatelessWidget {
         decoration: InputDecoration(
             hintText: this.hint,
             hintStyle:
-                Theme.of(context).textTheme.body1.copyWith(color: hintColor),
+                Theme.of(context).textTheme.body1?.copyWith(color: hintColor),
             errorText: errorText,
             counterText: '',
             icon: this.isIcon ? Icon(this.icon, color: iconColor) : null),
