@@ -10,18 +10,18 @@ class DioClient {
   // Get:-----------------------------------------------------------------------
   Future<dynamic> get(
       String uri, {
-        Map<String, dynamic> queryParameters,
-        Options options,
-        CancelToken cancelToken,
-        ProgressCallback onReceiveProgress,
+        Map<String, dynamic>? queryParameters,
+        Options? options,
+        CancelToken? cancelToken,
+        ProgressCallback? onReceiveProgress,
       }) async {
     try {
       final Response response = await _dio.get(
         uri,
-        queryParameters: queryParameters,
-        options: options,
-        cancelToken: cancelToken,
-        onReceiveProgress: onReceiveProgress,
+        queryParameters: (queryParameters)!,
+        options: (options)!,
+        cancelToken: (cancelToken)!,
+        onReceiveProgress: (onReceiveProgress)!,
       );
       return response.data;
     } catch (e) {
@@ -34,21 +34,21 @@ class DioClient {
   Future<dynamic> post(
       String uri, {
         data,
-        Map<String, dynamic> queryParameters,
-        Options options,
-        CancelToken cancelToken,
-        ProgressCallback onSendProgress,
-        ProgressCallback onReceiveProgress,
+        Map<String, dynamic>? queryParameters,
+        Options? options,
+        CancelToken? cancelToken,
+        ProgressCallback? onSendProgress,
+        ProgressCallback? onReceiveProgress,
       }) async {
     try {
       final Response response = await _dio.post(
         uri,
         data: data,
-        queryParameters: queryParameters,
-        options: options,
-        cancelToken: cancelToken,
-        onSendProgress: onSendProgress,
-        onReceiveProgress: onReceiveProgress,
+        queryParameters: (queryParameters)!,
+        options: (options)!,
+        cancelToken: (cancelToken)!,
+        onSendProgress: (onSendProgress)!,
+        onReceiveProgress: (onReceiveProgress)!,
       );
       return response.data;
     } catch (e) {
