@@ -13,7 +13,7 @@ class SharedPreferenceHelper {
   // General Methods: ----------------------------------------------------------
   Future<String> get authToken async {
     return _sharedPreference.then((preference) {
-      return preference.getString(Preferences.auth_token);
+      return preference.getString(Preferences.auth_token)!;
     });
   }
 
@@ -58,7 +58,7 @@ class SharedPreferenceHelper {
   // Language:---------------------------------------------------
   Future<String> get currentLanguage {
     return _sharedPreference.then((prefs) {
-      return prefs.getString(Preferences.current_language);
+      return prefs.getString(Preferences.current_language)!;
     });
   }
 
