@@ -5,10 +5,10 @@ class Post {
   String body;
 
   Post({
-    this.userId,
-    this.id,
-    this.title,
-    this.body,
+    required this.userId,
+    required this.id,
+    required this.title,
+    required this.body,
   });
 
   factory Post.fromMap(Map<String, dynamic> json) => Post(
