@@ -4,11 +4,11 @@ class PostList {
   final List<Post> posts;
 
   PostList({
-    this.posts,
+    required this.posts,
   });
 
   factory PostList.fromJson(List<dynamic> json) {
-    List<Post> posts = List<Post>();
+    List<Post> posts = <Post>[];
     posts = json.map((post) => Post.fromMap(post)).toList();
 
     return PostList(
